requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
Pillow==10.2.0
matplotlib==3.8.2
python-dotenv==1.0.0
Flask==3.0.0
requests==2.31.0
ollama==0.1.6
google-generativeai==0.3.2
diffusers==0.25.0
torch==2.1.2
transformers==4.36.2
accelerate==0.25.0
